Rename `plumbing.rs` as `incremental.rs`
Because it now contains only things related to incremental. Also move
two functions relating to incremental from `execution.rs` to
`incremental.rs`.
